Fix multi screen quick tiling
When a quick tiled window is moved between the screens, it will migrate
between the tile trees. The Tile::addWindow() will evacuate the window
from the previous tile only if both tiles belong to the same tree. It's
not going to be the same when moving a window between outputs. We need to
handle it explicitly.