backends/drm: always apply night light with the absolute colorimetric transform
The matrix optimzies down to the previously used channel factors (as long as no ICC profile is
involved), which means we ultimately get the same color operations out of it, but with slightly
less code.
To do that transform, this commit introduces the blending space of a drm output, which will also
be useful for overlay plane usage in the future.