x11window: don't wait for ack to make applet popups exposed immediately
The sync request often takes a significant amount of time (≥400ms) as Qt sends the
ack event using postEvent, which means a QtQuick window can only be exposed
until almost all animations are completed. While this is OK for
normal QtQuick windows as it hides the broken initial states in those
windows, it's problematic for applet popups as those popups usually has
a slide-in effect, which makes the user feel the system is very
unresponsive.

BUG: 504283
FIXED-IN: 6.4